using_the_python_interpreter.md - Alvis and Dardel docs now considered Good
This evaluation is based on my reading of the pages now linked (perhaps influenced by what I saw or experienced finding my way to these pages). It may be that deeper or wider browsing of the docs leads to a different evaluation, but that is not relevant for this part of this course.